requirements.txt
numpy==1.26.4
pandas==2.0.3
scikit-learn==1.2.2
packaging==25.0
tensorboard==2.14.0
torchdata==0.7.1
pyyaml==6.0.2
pydantic==2.10.6
tqdm==4.67.1
matplotlib==3.10.7